requirements.txt
# Core AWS & API
boto3>=1.26.137
botocore>=1.29.137
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Data Processing & ML (Python 3.13 compatible)
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional (for advanced ML)
sagemaker>=2.175.0
